refactor: move backend test mock data to data directory
